fix gl3 and redundant linking in bridge
