+++ ENIGMAsystem/SHELL/Bridges/Win32-Direct3D11/Makefile
@@ -1,2 +1,1 @@
 SOURCES += Bridges/General/Win32-Direct3D11.cpp $(wildcard Bridges/Win32-Direct3D11/*.cpp)
-override LDLIBS += -ldxgi -ld3d11
