+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3matrix.cpp
@@ -16,6 +16,7 @@
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
 
+#include "GL3matrix.h"
 #include "Bridges/General/GL3Context.h"
 #include "Graphics_Systems/General/OpenGLHeaders.h"
 #include "Graphics_Systems/General/GSmatrix.h"
